--- src/App.css
@@ -1,19 +1,0 @@
-*, html {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-}
-
-body {
-  font-family: 'Inter,-apple-system,BlinkMacSystemFont,"Segoe UI",Helvetica,Arial,sans-serif,"Apple Color Emoji","Segoe UI Emoji"';
-  font-size: 62.5%;    /* 62.5% = 10px */
-}
-
-:root {
-  --font-body: 'Inter,-apple-system,BlinkMacSystemFont,"Segoe UI",Helvetica,Arial,sans-serif,"Apple Color Emoji","Segoe UI Emoji"';
-  --font-headline: 'Inter,-apple-system,BlinkMacSystemFont,"Segoe UI",Helvetica,Arial,sans-serif,"Apple Color Emoji","Segoe UI Emoji"';
-  --white-bg: #fff;
-  --input-border: 1px solid rgb(208, 212, 220);
-  --input-border-focus: 1px solid rgb(59, 60, 62);
-  --input-txt-color: rgb(103, 112, 132);
-}
