+++ src/App.js
@@ -1,12 +1,11 @@
-import './App.css';
-import { Route, Routes } from 'react-router-dom';
-import Login from './pages/Login';
-
+import { Route, Routes } from "react-router-dom";
+import Login from "./pages/Login";
+import "./sass/main.scss";
 function App() {
   return (
     <div className="App">
       <Routes>
-        <Route exact path='/' element={<Login title="Log in | Contra" />} />
+        <Route exact path="/" element={<Login title="Log in | Contra" />} />
       </Routes>
     </div>
   );
