+++ src/sass/abstracts/_variables.scss
@@ -1,0 +1,16 @@
+// fonts
+
+$font-body: 'Inter,-apple-system,BlinkMacSystemFont,"Segoe UI",Helvetica,Arial,sans-serif,"Apple Color Emoji","Segoe UI Emoji"';
+$font-headline: 'Inter,-apple-system,BlinkMacSystemFont,"Segoe UI",Helvetica,Arial,sans-serif,"Apple Color Emoji","Segoe UI Emoji"';
+
+
+// colors
+
+$color-white: #fff;
+$color-black: #000;
+
+// borders
+
+$input-border: 1px solid rgb(208, 212, 220);
+$input-border-focus: 1px solid rgb(59, 60, 62);
+$input-txt-color: rgb(103, 112, 132);
