+++ src/sass/base/_base.scss
@@ -1,0 +1,16 @@
+*,
+*::after,
+*::before {
+    margin: 0;
+    padding: 0;
+    box-sizing: border-box;
+  }
+  
+html {
+    font-family: 'Inter,-apple-system,BlinkMacSystemFont,"Segoe UI",Helvetica,Arial,sans-serif,"Apple Color Emoji","Segoe UI Emoji"';
+    font-size: 62.5%;    /* 62.5% = 10px */
+}
+  
+body {
+    box-sizing: border-box;
+}
