--- src/styles/Login.css
@@ -1,184 +1,0 @@
-.login {
-    height: 100vh;
-    flex: none;
-    box-sizing: border-box;
-    position: relative;
-}
-
-.login__bg > img {
-    width: 100%;
-    height: 100vh;
-    background-repeat: no-repeat;
-    background-size: cover;
-    filter: blur(20px);
-}
-
-/* logo position */
-
-.login__banner {
-    position: absolute;
-    top: 3em;
-    left: 5em;
-}
-
-/* svg {
-    filter: invert(1);
-} */
-
-/* container */
-
-.login__page {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    position: relative;
-}
-
-.login__container {
-    box-sizing: border-box;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    position: absolute;
-    bottom: 15em;
-    right: 24em;
-    background: var(--white-bg);
-    border-radius: 1.2rem;
-    width: 70%;
-    height: 70vh;
-    padding-left: 6em;
-}
-
-.login__container-left {
-    flex: 0 0 50%;
-    height: 100%;
-    display: flex;
-    flex-direction: column;
-    justify-content: flex-start;
-    align-items: flex-start;
-    text-align: left;
-    margin-top: 33%;
-}
-
-.login__container-left--title {
-    font-size: 2.125rem;
-    font-weight: 700;
-    letter-spacing: -0.5px;
-    line-height: 40px;
-    margin-bottom: 1em;
-}
-
-.login__container-left--input-cont {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: flex-start;
-    position: relative;
-    width: 90%;
-}
-
-.login__container-left--input-cont > label {
-    position: absolute;
-    top: .5rem;
-    left: 1.45rem;
-    font-size: 0.75rem;
-    font-weight: 400;
-    line-height: 1.333;
-    color: var(--input-txt-color);
-}
-
-input[type="text"] {
-    padding-top: 1.9rem;
-    padding-left: 1.3rem;
-    padding-bottom: .3rem;
-    width: 100%;
-    border-radius: .6rem;
-    border: var(--input-border);
-    transition: all .4s ease-in-out;
-    margin-bottom: 2em;
-}
-
-input[type="text"]:focus {
-    padding-top: 2.7rem;
-    padding-bottom: .8rem;
-    outline: none;
-    border: var(--input-border-focus);
-    transition: all .4s ease-in-out;
-}
-
-input[type="text"]::placeholder {
-    position: absolute;
-}
-
-
-/* right side */
-
-.login__container-right {
-    height: 100%;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-.login__container-right > img {
-    width: 90%;
-}
-
-
-/* btn login */
-
-.login__container-left--btn {
-    padding: .8rem 1.1rem;
-    background-color: black;
-    border: none;
-    color: var(--white-bg);
-    border-radius: 1.5rem;
-    font-weight: 600;
-    font-size: 0.875rem;
-    line-height: 1.42;
-    cursor: pointer;
-}
-
-
-/* sign up container */
-
-.signup__container {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    position: absolute;
-    bottom: 8em;
-    gap: 1rem;
-}
-
-
-.signup__container-text {
-    color: black;
-    max-width: 480px;
-    text-align: center;
-    font-weight: 400;
-    font-size: 1.2rem;
-    letter-spacing: 0.15px;
-    line-height: 1.42;
-}
-
-.signup__container-btn {
-    padding: .8rem 1.1rem;
-    background-color: black;
-    border: none;
-    color: var(--white-bg);
-    border-radius: 1.5rem;
-    font-weight: 600;
-    font-size: 0.875rem;
-    line-height: 1.42;
-    cursor: pointer;
-    transition: all .3s ease-in-out;
-
-}
-
-.signup__container-btn:hover {
-    color: black;
-    background-color: white;
-    transition: all .3s ease-in-out;
-}
